Fix Bugs on linux64

+++ src/ofd/Signature.cc
@@ -16,13 +16,13 @@ void test_libsodium(){
     crypto_sign_keypair(pk, sk);
 
     uint8_t signedMessage[crypto_sign_BYTES + msgSize];
-    uint64_t signedMessageLen = 0;
+    unsigned long long signedMessageLen = 0;
 
     crypto_sign(signedMessage, &signedMessageLen, (const uint8_t *)msg, msgSize, sk);
 
 
     uint8_t unsignedMessage[msgSize+1];
-    uint64_t unsignedMessageLen = 0;
+    unsigned long long unsignedMessageLen = 0;
     if ( crypto_sign_open(unsignedMessage, &unsignedMessageLen, 
                 signedMessage, signedMessageLen, pk) != 0 ){
         LOG(ERROR) << "Message sign is wrong.";
